requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.3
asyncpg==0.29.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
coverage==7.6.7
cryptography==43.0.0
cssselect2==0.7.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.112.1
fonttools==4.55.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.7
iniconfig==2.0.0
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.2
passlib==1.7.4
pdfkit==1.0.0
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.15.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-order==1.3.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.1.0
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.30.6
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
